refactor: migrate compose utilities to TypeScript
