+++ src/legacy-lib/js-yaml.d.ts
@@ -1,0 +1,3 @@
+declare module 'js-yaml' {
+  export function load(input: string): unknown;
+}
